Fix buffer memory reporting
We can't use 'uint64_t' here, because it's mapped to 'ulong' Qt type,
which is not registered in Qt DBus type system. On the other hand,
'unsigned long long' is okay, because it maps to 'qulonglong', which
maps to DBus type 'uint64', which is what we want.
See https://doc.qt.io/qt-6/qdbustypesystem.html for details.

BUG: 498373

+++ plugins/memory/freebsdbackend.cpp
@@ -58,7 +58,8 @@ void FreeBsdMemoryBackend::makeSensors()
     m_cache = cacheSensor;
     m_sysctlSensors.push_back(cacheSensor);
 
-    auto bufferSensor = new KSysGuard::SysctlSensor<uint64_t>(QStringLiteral("buffer"), "vfs.bufspace", m_physicalObject);
+    // uint64_t can't be used here, because it's mapped to Qt type ulong, which is not registered in Qt DBus type system. See https://doc.qt.io/qt-6/qdbustypesystem.html
+    auto bufferSensor = new KSysGuard::SysctlSensor<unsigned long long>(QStringLiteral("buffer"), "vfs.bufspace", m_physicalObject);
     m_buffer = bufferSensor;
     m_sysctlSensors.push_back(bufferSensor);
